Move models test outside of /api directory
